Bug fixes and improvements
* Added a SHIFT key modifier to the QuestTracker and QuestArrow - unless

the shift key is down, you can't drag it.

* Wardz's commit fixing/preventing the QuestTracker from being dragged
off screen was also applied to the QuestArrow.

* Added Enable/Disable toggles for
- Worldmap/Minimap icons (/questie mapaids)
- QuestArrow (/questie arrow)

* Added Bronze Feather to the Winfury Matriarch drop table for the quest

Preparation for Ceremony reported by PierrickGT.

* Added a small Worldmap modifier for those not running a Mapmod like
Cartographer or MetaMap. This allows the user to see the chat output
and allows the Popup dialog to appear so they can select a quest step
to manually complete.

* I discovered a typo in the code pertaining to EQL3. This fix now
allows
the user to untrack a quest from their QuestTracker with or without
EQL3.