[Bug] Fixed a nil error for EQL3 users
On certain servers EQL3 would inject the quest [%level%] tag into the
Quest Title. When a player hands in a quest they get a nil error on
servers that block this behavior or if the player is using a version
that doesn't add the quest level. Or possibly the player has this
feature turned off. This check handles all scenarios.