[Hotfix] Worldmap and Quest management issues
* Hotfix: fixed an issue with Worldmap code for users that aren't using
a map mod. Added a couple of new nil checks that were popping up for
some users who recently installed Questie and don't have any Saved
Variables file at all. I thought I had thoroughly checked this code but
that was before the global to local conversion and code rewrite. Sorry
guys!

* Hotfix: fixed an issue with one of the quest management functions. I
had
to add some additional nil checks to make sure a quest is actually
present
in the database before setting the saved variables data keys in the db.
If a quest isn't present, then it will now silently ignore the nil error
but the quest can't be tracked. Questie's quest database was parsed from
a well-known and trusted source based on quest data obtained from the
live
retail realms in the very early days of BC. Some of the Vanilla quest
data
might be missing or slightly inaccurate. We're in the process of
creating
a new database for both the Vanilla and TBC clients of Questie and all
sources will be scrapped and checked for duplicate entries from both the
current database and either the most recent mangos distro or some other
well respected project.