[Hotfix] Astrolabe - Darnassus fix
About a month ago someone submitted more accurate zone size data for
Astrolabe. The fix was merged with Questie but unfortunately it broke
the city of Darnassus and caused all the Minimap icons to appear over
the player arrow indicator on the Minimap. The quest NPC's and
objectives were no longer stationary in relation to where they were
supposed to be positioned. This also broke the QuestieArrow. The old
values were imported from the old Astrolabe file and the issue has been
resolved.