[Hotfix] Astrolabe - Zone Spellings
* Hotfix: The previous hotfix was incorrect in the assumption that the
updated data was incorrect. The city of Darnassis was spelled
incorrectly.
Someone reported issues with Ogrimmar today and I found that it too was
miss-spelled. I've gone through and checked the reset of the cities to
make
sure they were all correct. I've also reverted the city of Darnassis
dimensions back to the submitted values from a month ago.