User Submitted Code Change
This allows Quest items to display their tooltip without expanding a
"collapased/minimized" section in the quest log. In the Default WoW
Quest Log you can collapse a header which is usally a list of quests in
a zone, area or catagory. Previously, tooltips wouldn't show up unless
the section the quest was contained in was expanded.