[Hot Fix] Quest Objectives Flickering and Overall Weirdness
* I think I may have traced the issue. In Vanilla WoW the only way to
obtain
a player’s true position is via the Worldmap. When you open the map, and
start
browsing other zones, your Avatars position coordinates change
accordingly.
Questie relies heavily on Astrolabe to determine players position and
distances
to quest objectives. This can cause weird issues with the QuestArrow,
QuestTracker, and the MapNotes that are drawn on the Minimap and
Worldmap.
I've added some basic logic code to determine which zone the player is
physically in and if the world map is open, all distance calculations
are
halted under certain circumstances.
* The code that draws the map icons has also been streamlined, sped up
and
tweaked in such a way that they now refresh much quicker and ONLY when
absolutely needed. This should solve the "flickering" issue that some
users
have reported.
* Since I was in Astrolabe and updating it anyway, I've added a small
feature.
This feature not being present has been bugging me for quite a while and
when someone mentioned TomTom the other day, it reminded me that it's a
feature that I've always wanted. If you die, Questie will paint an arrow
to your Corpse automatically. When you resurrect, it'll disappear.