[Release] Version 3.65
* Added a new feature that allows you to turn off the Questie tool tips.
Use '/questie tooltips' to toggle them on and off. They are on by
default.
* Fixed some issues with tool tips causing all kinds of weird issues,
everything from duplicate tool tips to nil errors.
* Fixed the minlevel and maxlevel functions. One of the new variables
wasn't set correctly when the new Required Level field was added to the
quest database. These filters are now keyed off of the Required Level
instead of Quest Level so don't be surprised if you see a level 60 quest
appearing on your map. Chances are it's one of those quests that can be
done at level 5. Like some of the holiday quests.
* I've also added a new line to the Questie MapNotes tooltips. In
addition
to [QuestLevel][QuestTitle] followed by [QuestStarter} etc. I've
inserted
[Min Level] info, or what was once referred to as Suggested Level in
Vanilla.
By default, the minlevel filters hides quests 3 levels BELOW your
current
level based on the Minimum Level of the Quest. For example, if a quests
Level is 40 and its Min Level is 35 and your character’s level is 39
then this quest will be hidden if minlevel is on. You can change the
number of levels it filters higher or lower. Default is 3. Same example,
with maxlevel this quest WON'T show until your character reaches a level
of 30. Default maxlevel filter is 5 which means quests won't show until
you are within 5 levels of the quests Min Level.