Added UI support for settings
Added /questie options to display the newly added user interface for
managing settings. All the old commands still work but now there is an
options for those who would like a simple interface for changing their
settings. The UI file is compatible with WoW UI Designer for easy
updating in the future.