Added warning message for options that require a reload
If the warning to reload the ui is not accepted, the settings will be
reverted to before the user attempted to change them. A chat message is
displayed when this happens. If the warning is accepted the UI will be
reloaded.

All settings that require a reload are enumerated in the error message.