Quest Management Fixes
*Zoey and I found the issue causing her game client to freeze. This
threw off several key functions firing sequences and allowed me to
remove a ton of redundant calls. I also balanced the timings so that
players on above average latency connections (400ms+); won’t miss any
Quest Events even if their QuestLog has the max number of quests and the
local game client cache is empty.
*Reverted Questie:Toggle() back to its original state. The basic Toggle
only turns off the MapNotes and does not prevent Questie from continuing
to collect and cache Quest data. In the event that the user chooses to
switch back on the Toggle, all quests and markers will be up to date.
It’s been suggested that we conduct a user poll to see if users want the
Toggle to become an in game “On/Off” switch. The only caveat to that is
that Questie will not update the Quest Cache and information can become
out of sync. When the user turns the Toggle back on, Questie will have
to perform a series of QuestLog checks similar to Player Logon to bring
all Quest data back into sync. My suggestion is to only switch off the
Tracker and MapNotes but allow Questie to continue running in the back
ground and updating Quest Caches.
*Added a check to the Questie:CheckQuestLog function that runs upon
Logon and ReloadUI to check for Quests flagged as active (0) in
QuestieSeenQuests and the QuestieCachedQuests cache (if applicable) that
isn’t active in the players QuestLog. This can cause nil errors further
down the workflow if it’s trying to find a hash value for a quest that
doesn’t exist in other caches.
*Fixed an issue with the TrackerHeader extending beyond the back-ground
border. I added an auto-width check that if any quest(s) is/are narrower
than the header, to use the headers width instead.