Quest Management Fixes
* Quest Management: Reverted most of the timings that were making
Questie appear sluggish
* Quest Management: Cleaned up the initialization workflows and sped up
some of the timings
* Quest Management: Added some additional checks to
UpdateGameClientCache() to make sure cached data is accurate and not
garbling the QuestieCachedQuests table.
* Auto Finish Quests – Inventory Check: Tried GetNumQuestLogRewards()
but it was returning zero for some strange reason so I just made a < 1
check for avail slots rather than a hard == 0.
* Auto Finish Quests – Inventory Check: Removed inventory check for
Quest Progress finishers
* Quest Items: Cleaned up and fixed an issue with the Detect Quest Items
* EQL3 Support: Added an Anti-Event flooder for EQL3 – it was causing my
client to freeze. Did some research and found a fix that was written by
Shagu. Thanks for the code Shagu!
* Default WoW Quest Log: Removed the initWOWQuestLog after re-writing
syncWOWQuestLog and renaming it to syncQuestWatch – it’s now a true
QuestWatch sync and doesn’t interfere with EQL3.
* Quest Management: Improved syncQuestLog – before it was calling
addQuestToTracker, addQuestToTrackerCache and AddQuestToMap each time a
quest was added to the tracker. This caused a huge memory/lag spike when
adding quests in bulk to the tracker. Now it checks to see if the quest
is cached in QuestieCachedQuests. If a quest is cached it calls
addQuestToTracker, if it’s not cached then it calls all three.