Re-added "Time to Arrive" on QuestieArrow
I've re-added a feature that was part of Questie a long time ago.  Long story short we thought it was too memory intensive but it turns out that drawing the arrow was the culpret. I just never got around to re-adding the Time to Arrive counter back in. It's off by default. To turn it on, just run the "/questie arrowtime" command.