QuestieTBC: Fix tooltip-breaking call to gfind
In TBC, the "string.gfind" function has been removed from the game
engine. Trying to call it causes a Lua error.

And since the "gfind" call was crashing inside the tooltip handler,
it meant that whenever the "gfind" code ran, it broke the in-game
tooltip completely, making it invisible until the person logs out again.

This bug was probably missed during the previous TBC porting simply
because the tooltip scanning code in Questie doesn't always run. So
this is a rare, but very serious bug.

However, after analyzing what the Questie code actually tried to do with
"gfind", I see that it was just scanning for the first match of the word
"objective" within the string "name". So I replaced it with a regular
"string.find" instead of a "gfind". I also added comments to explain
what it's doing, since the variable-names themselves made no sense.

You can read about the different WoW string search methods here:
http://wowwiki.wikia.com/wiki/Pattern_matching#.22Regexes.22_.28patterns.29_in_Lua