QuestieTBC: Fix quest-linking via quest log frame
Questie was being very naughty. It completely replaced Blizzard's
built-in "QuestLogTitleButton_OnClick" function, and instead ran
some outdated code which simply inserted a text-only description
of the quest.

This problem meant that it was impossible to shift-click on quests
to insert proper, clickable quest-links while using Questie in TBC.

It has now been fixed, so that it runs the Blizzard click-handler
first and *then* does the Questie-specific extra actions (toggling
questie tracking and hiding the default Blizzard watcher frame).