QuestieTBC: Fix tooltip error for missing quests
Certain quests can end up on the worldmap even though Questie doesn't
know about them, as long as Questie knows its "finisher NPC's" location.

This happened with the TBC quest "Contact from Dalaran" (hash
3003607748), which lacks hash-info but is marked in QuestieTBC's database:

Database/finishers.lua
2399:  ["Contact from Dalaran"] = "Archmage Cedric",

Which placed the "?" icon on the world map at the correct location.
But actually hovering over the icon produced a Lua error since the
"QuestieHashMap" doesn't contain the questID/hash.

We now verify the existence of the quest info first, and if not, we show
a debug tooltip with the quest name and ID instead of throwing an error!