Cleanup db, change how class/race prune works
The reason for this change is, the race / class requirement data has been wrong in a couple cases. This way it only hides available quests, if the player has the quest in their log already it will still add objectives this way. Also added the grave spawn data, and pruned the special requirements table to remove redundant data