Commence search refactoring
Result detail frames are still WIP. The quest one is missing some
features from the old one, the npc frame is the same, the other two are
missing completely

* Added new search result page based on tabs.
* Added new `QuestieSearch` class, for doing the actual searching and
returning a table of results from all categories. This result table is
also available as `QuestieSearch.LastResult`.
* Added `npcKeys` and `itemKeys` tables until they can be properly added
to the DB files.
* Miscellaneous formatting fixes.