Fix search checkbox for hidden quests
The checkbox wasn't initialised correctly, which lead to it showing the
wrong state when a quest was actually hidden. Now it is initialised, but
there is still a desync when the quest is hidden by Shift-Click while a
search is open. This is fixed by loading the search result page again
though, so not a big issue.

fixes #1000