Fix quests being "abandoned" when a quest is turned in
The QUEST_REMOVED event was marking a quest as abandoned regardless of
being fired for an abandoned or finished quest. To correct this the
quest ID is saved on QUEST_TURNED_IN event and then compared on
QUEST_REMOVE. If it is the same, marking the quest as abandoned is
skipped.

This fixes #1238