[quest] Fixes for Sunken Treasure
There was an issue a while back where someone asked to have the gears made transparent so they could see these items. I had responded... "If you zoom in you can see the active spots vs. just gears..." Anyway that was dumb.

This is better, just rounding the locations of the spawn points so that we still see the general area, but the gears and the actual gems won't perfectly overlap. Better solution.

* Calcified Elven Gem - Object - World of Warcraft || https://classic.wowhead.com/object=2712/calcified-elven-gem
* Sunken Treasure - Quest - World of Warcraft || https://classic.wowhead.com/quest=666/sunken-treasure