Taste the Rainbow
Hey I made some tweaks before, but this is the final one. I didn't realize the list was random. I was trying to get the colors to sort in a specific order that wouldn't overlap. It's sorted that way now, should cycle through about every 6 colors it'll load a "blue-ish" but similar icons shouldn't touch.