[quest] Fix objective locations for "Skorn Must Fall!"
I'm not happy about this fix, since it only adds a location for the quest recipient, but it will do for now. Ideally, we should get rid of this recipient location data (keep it nil) and place an extra objective showing where to use the quest item (33340) to spawn the recipient NPC instead. Unfortunately this marker won't ever show, because the quest is considered completed as soon as you pick it up. At least now, you know where to go.