[quest] Fix objective locations for "Not a Bug" (all)
Fixes #4333 more thoroughly.

Ideally, we would actually point the Objectives at the object for "Summoning Stone", but this object does not seem to exist in our database. This may be fixed in a followup commit soon. Until we have an object ID, the Summoning Stone will not have a tooltip.