[quest] Fix coordinates for Stormwind Candy Bucket
These coordinates were wrong due to us loading TBC corrections on top of WotLK data, and TBC lacked the harbor so the coordinates are all different. While deleting the TBC correction seems a bit heavy-handed, the correction was only for a difference of 0.01 on the Y axis, so even if TBC exists again and Questie gets used in it, nobody will notice the difference here.