[quest] Fix objective locations for "Forging a Head"
These fixes are a bit convoluted. We previously attributed the item drops directly to the dead iron giants, but that's not entirely how this quest actually behaves, so I added them to the mob that actually drops them (30208) and added extraObjectives to help indicate how to progress the objective. Also updated with corrected locations for 30163 from Wowhead, and erased the 30208 spawns since those now being exposed can cause confusion. I also removed references to 29914 since that mob does not seem to exist in the phase this quest takes place in.