[quest] Fix a TBC-era regression where some class-specific starting zone quests show as available to other races
This shows up as an issue in both TBC and Wrath, but is not an issue in Classic since our Classic database is healthy. Fixing this in the TBC corrections since it affects the TBC database (just incase there's ever another TBC release), and will get included in Wrath automatically.

Fixes #4372