[quest] Fixed lots of Love is in the Air quests
For "Crushing the Crown" quests, all of them seemingly targeted a single kill bunny NPC, which we only had spawn points for in Crystalsong. To avoid confusion, I removed the NPC's spawn points, and then added dummy objects that we can add spawn points to for the various zones. I also hooked the quests up to these objects with extraObjectives.