QA Team Requests and Fixes
Quest Item Buttons:
1. Fix charges/stack size indicator - Not showing. (FIXED)
2. Change font for charges/stack size. (FIXED)
3. Verify that Quest Item button disappears if player runs out of items in inventory. (FIXED)

Questie Tracker:
Auto width is still truncating quests. (FIXED)

Questie Tracker Options:
1. Re-add Minimize Completed Quests option. (FIXED)
2. Add separate Hide Completed Quest/Achievement Objective(s) option. (FIXED)
3. Test Questie Options and Questie Journey in and out of combat. Prevent it from opening while in combat and close it if open upon entering combat. (FIXED)

Not able to fix:
1. Check for flag in QuestDB for max number of quest items stack. (NONE USABLE)
2. Look into flipping Quest Item button for "multi-step" Quest Item buttons (example Way of the Heart - flip between Net and Fish)

This feature just isn't currently possible. There is no way to check a quest for "if have min number of dropped items to complete quest". I was thinking about keying off of the 'Max Number of Items Per Stack' which would work for "Hot and Cold" (Max Stack Size: 6) but not "The Way to His Heart..." (Max Stack Size: 20). A player just won't farm up that many of them if they only need 5-10. This type of feature is what @Logon was talking about a few days ago about Quest Conditionals. You all could manually add a "Conditional" field/key in the QuestDB like 'questItemsNeeded' but then you'd all have to manually add those in Questie QuestFixes right? Do you really want to do that? Even if I hard coded a set value (it would have to be a minimum of 6), but I don't know if it's enough. Some quests might need more than 6 drops to complete like in "The Way to His Heart...". If a player was good and understood the mechanics then yeah, I'm sure 6 fish would be enough. But not always...