Added Secondary Quest Item Buttons!
Not much to describe. If you're on a quest and an NPC gives you an item that you use to gather other items needed to complete an objective; a secondary quest item button will appear when those items are looted from mobs or picked up in the world. For example, The Way to His Heart... in Howling Fjord. The primary Quest Item Button will be a net. When you use the net to catch the fish then a secondary Quest Item Button will appear allowing you to use the fish.