Reverted to an older method I used to insert spaces between tracker elements.
The new method put actual spaces between objects which made the onHighlight feature act weird because it would flash on and off while the mouse transitioned down the list of Quests/Achievements. The old method increased the line height simulating a space resulting in no physical space between elements which allowed the onHighlight feature to transition between elements of the same Quest/Achievement smoothly without flashing on and off.