Turned off achievement event: CRITERIA_UPDATE
This fires FAR too often. Until Blizzard figures out a way to allow us to trigger achievement updates this needs to remain disabled for now.

In the meantime I'm going to do some research and see if there is a way that I can filter the event so that it only fires an update when an actual achievement needs updating.

NOTE TO SELF: Consolidate my current method under the CRITERIA_UPDATE event and use those as a filter so they are all in one place in the code base vs being inserted into other events.