Quest Item Button was needed after Quest was completed
I came upon an issue with Quest Item buttons. There was an item that was usable after a quest was complete and it disappeared. In order to get all quests to work, the Primary quest item buttons should always be shown even after a quest is completed unless, "Blizzard Completion" text returns true. I have yet to encounter a quest where a quest item was needed when the quest reaches this state.