[Fix] Fixed "Sizer Mode: Manual" window stuck to pointer
Fixes #4916

Not sure why it's getting stuck. We've been using the same Tooltip code for quite sometime. This is the first I've heard of it. However, the Tracker Tooltips were anchored to UIPARENT and the Map Icon Tooltips are anchored to their own frames. Just in case this is some weird edge case I've changed the anchor method for the Tracker Tooltips to anchor to itself vs UIPARENT.