[fix] fixes the map/minimap tooltips where the text was split on dot
The map/minimap tooltips have no reason to split on dot. There are NPCs with "." in their name and the way it was initially was leading to some weird and unnecessary text formatting inside the text definitions from the DB. The outcome is much cleaner like this.
Furthermore, it is splitting on dot no matter the value (true or false, it still splits), and it was forcefully splitting the line regardless.
Fixes #4892