[fix] Fix for very short Blizzard Completion text
Normally, Blizzard Completion text is quite verbose in nature so we assign a "trackerMinLineWidth" of 280 pixels because anything less, the wrapped text looks stupidly squished. In this case the text is quite short and results in a ton of "empty space". Added a "catch all" check where we compare the Quest Title and the wrapped text length. If both are short, then use the shorter measurement else, make the line a min of 280 pixels or if it's longer, use the longer one.