[fix] Fix for very short Blizzard Completion text
Normally, Blizzard Completion text is quite verbose in nature so we assign a "trackerMinLineWidth" of 280 pixels because anything less, the wrapped text looks stupidly squished. In this case the text is quite short and results in a ton of "empty space". I had to go back and add a check in the Tracker Update to not wrap "short" lines specifically.