[fix] Fix for attempt to index upvalue 'trackerBaseFrame' (a nil value)
Happened when a player died and came back to life... evoked from the PLAYER_ALIVE event.

These upvalues don't need to be set unless the Tracker is visible on the screen or has a tracked quest or not disabled or not minimized or not fully initialized. And obviously not having the addon VoiceOver enabled.